% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_scales.R
\name{scale_color_wixarika}
\alias{scale_color_wixarika}
\alias{scale_colour_wixarika}
\alias{scale_fill_wixarika}
\title{Wixarika Color Scales for ggplot2}
\usage{
scale_color_wixarika(name, discrete = TRUE, reverse = FALSE, ...)

scale_colour_wixarika(name, discrete = TRUE, reverse = FALSE, ...)

scale_fill_wixarika(name, discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{name}{Name of the palette}

\item{discrete}{Logical. Use discrete (TRUE) or continuous (FALSE) scale}

\item{reverse}{Logical. Reverse the palette?}

\item{...}{Additional arguments passed to ggplot2 scale functions}
}
\description{
Wixarika Color Scales for ggplot2
}
\examples{
\dontrun{
library(ggplot2)

# Discrete color scale
ggplot(iris, aes(Sepal.Length, Sepal.Width, color = Species)) +
  geom_point(size = 3) +
  scale_color_wixarika("muvieri_vivid")

# Continuous fill scale
ggplot(faithfuld, aes(waiting, eruptions, fill = density)) +
  geom_tile() +
  scale_fill_wixarika("hikuri_seq", discrete = FALSE)
}
}
