% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{wixarika_pal}
\alias{wixarika_pal}
\title{Get a Wixarika Palette}
\usage{
wixarika_pal(name, n = NULL, type = "discrete", reverse = FALSE)
}
\arguments{
\item{name}{Name of the palette. Use \code{wixarika_pal_names()} to see available palettes.}

\item{n}{Number of colors to return. If NULL, returns all colors in the palette.}

\item{type}{Type of palette: "discrete" or "continuous"}

\item{reverse}{Logical. Should the palette be reversed?}
}
\value{
A character vector of hex color codes
}
\description{
Get a Wixarika Palette
}
\examples{
# Get a complete palette
wixarika_pal("tatewari_warm")

# Get 3 colors from a palette
wixarika_pal("kauyumari_forestsky", n = 3)

# Reverse a palette
wixarika_pal("tayau_dawn", reverse = TRUE)
}
