% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{wixarika_pal_names}
\alias{wixarika_pal_names}
\title{List Available Wixarika Palettes}
\usage{
wixarika_pal_names()
}
\value{
A character vector of palette names
}
\description{
List Available Wixarika Palettes
}
\examples{
wixarika_pal_names()
}
