% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{wixarika_palettes}
\alias{wixarika_palettes}
\alias{wixarika_palette_info}
\title{WixarikaRt Color Palettes}
\format{
A list of color palettes

An object of class \code{list} of length 19.
}
\usage{
wixarika_palettes

wixarika_palette_info
}
\description{
A collection of color palettes inspired by the art and culture of the
Wixarika (Huichol) people and other Indigenous peoples of Mexico.
}
\details{
The palettes are inspired by:
\itemize{
\item Tatewari: Grandfather fire, sacred to the Wixarika people
\item Kauyumari: The sacred deer spirit
\item Tayau: Father Sun
\item Nakawe: Grandmother Earth and ocean
\item Tsikuri: Traditional ceremonial gourd bowls
\item Nierika: Portal to the spirit world
\item Muvieri: Arrows and ceremonial objects
\item Hikuri: Sacred peyote
\item Haramara: The Pacific Ocean
\item Kiekari: Sky and celestial elements
\item Mapa Lenguas: Pre-Hispanic indigenous languages distribution map
}
}
\examples{
# View available palettes
names(wixarika_palettes)

# Access a specific palette
wixarika_palettes$tatewari_warm

}
\keyword{datasets}
