% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{wixarika_show_all}
\alias{wixarika_show_all}
\title{Display All Wixarika Palettes}
\usage{
wixarika_show_all()
}
\value{
NULL (displays a plot)
}
\description{
Display All Wixarika Palettes
}
\examples{
wixarika_show_all()
}
